criação das chamadas para metodos de obter ocupação,calcular receita e verificações de existencias de room

+++ src/main/java/br/com/ada/reservala/service/RoomService.java
@@ -1,21 +1,27 @@
 package br.com.ada.reservala.service;
 
 import br.com.ada.reservala.domain.Room;
+import br.com.ada.reservala.exception.RoomNotFoundException;
 import br.com.ada.reservala.repository.RoomRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
+
 
 @Service
 public class RoomService {
 
-    private RoomRepository roomRepository;
+    private final RoomRepository roomRepository;
 
     public RoomService(RoomRepository roomRepository){
         this.roomRepository = roomRepository;
     }
 
     public Room createRoom(Room room){
+//        if (roomRepository.existsRoomNumber(room.getRoomNumber()) > 0){
+//            throw new RoomNotFoundException("Room with number " + room.getRoomNumber() + " already exists.");
+//        }
         return roomRepository.createRoom(room);
     }
 
@@ -23,22 +29,31 @@ public class RoomService {
         return roomRepository.readRoom();
     }
 
-    public Room updateRoom(Room room){
-        return roomRepository.updateRoom(room);
+    public Optional<Room> updateRoom(Room room, Integer roomNumber) {
+        if (roomRepository.existsRoomNumber(roomNumber) == 0){
+            throw new RoomNotFoundException("Room with number " + roomNumber + " not found.");
+        }
+        return Optional.of(roomRepository.updateRoom(room,roomNumber));
     }
 
     public void deleteRoom(Integer roomNumber){
+        if (roomRepository.existsRoomNumber(roomNumber) == 0){
+            throw new RoomNotFoundException("Room with number " + roomNumber + " not found.");
+        }
         roomRepository.deleteRoom(roomNumber);
     }
 
-    //Deve calcular o percentual de quartos ocupados
     public Double getOcupation(){
-        return 100d;
+        double allRoms = roomRepository.count();
+        if(allRoms == 0){
+            return 0.0;
+        }
+        double ocupiedRooms = roomRepository.countOccupied();
+        return (ocupiedRooms / allRoms) * 100;
     }
 
-    //Deve calcular a receita obtida
     public Double getRevenue(){
-        return 100d;
+        return roomRepository.countRevenue();
     }
 
 }
